Move add new collection into move book to collection dialog bottom sheet
